pyproject.toml
[tool.poetry]
name = "teia-logos"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "teia_logos"}]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
